Move integration test to composite builds subproject
The integration test which checks the behavior of configuration attributes resolution in
the context of composite builds has been moved to the composite builds subproject. This
removes the need for adding a dependency on the composite builds runtime to the integration
test runtime of the dependency management subproject.
